feat(provider/pihole): drop Pi-hole v5 API support
