+++ provider/pihole/pihole.go
@@ -22,7 +22,6 @@ import (
 	"slices"
 
 	"github.com/google/go-cmp/cmp"
-	log "github.com/sirupsen/logrus"
 
 	"sigs.k8s.io/external-dns/pkg/apis/externaldns"
 
@@ -35,15 +34,10 @@ import (
 // in the environment.
 var ErrNoPiholeServer = errors.New("no pihole server found in the environment or flags")
 
-const (
-	warningMsg = "Pi-hole v5 API support is deprecated. Set --pihole-api-version=\"6\" to use the Pi-hole v6 API. The v5 API will be removed in a future release."
-)
-
 // PiholeProvider is an implementation of Provider for Pi-hole Local DNS.
 type PiholeProvider struct {
 	provider.BaseProvider
-	api        piholeAPI
-	apiVersion string
+	api piholeAPI
 }
 
 // PiholeConfig is used for configuring a PiholeProvider.
@@ -58,8 +52,6 @@ type PiholeConfig struct {
 	DomainFilter *endpoint.DomainFilter
 	// Do nothing and log what would have changed to stdout.
 	DryRun bool
-	// PiHole API version =<5 or >=6, default is 5
-	APIVersion string
 }
 
 // Helper struct for de-duping DNS entry updates.
@@ -77,26 +69,17 @@ func New(_ context.Context, cfg *externaldns.Config, domainFilter *endpoint.Doma
 			TLSInsecureSkipVerify: cfg.PiholeTLSInsecureSkipVerify,
 			DomainFilter:          domainFilter,
 			DryRun:                cfg.DryRun,
-			APIVersion:            cfg.PiholeApiVersion,
 		},
 	)
 }
 
 // newProvider initializes a new Pi-hole Local DNS based Provider.
 func newProvider(cfg PiholeConfig) (*PiholeProvider, error) {
-	var api piholeAPI
-	var err error
-	switch cfg.APIVersion {
-	case "6":
-		api, err = newPiholeClientV6(cfg)
-	default:
-		log.Warn(warningMsg)
-		api, err = newPiholeClient(cfg)
-	}
+	api, err := newPiholeClientV6(cfg)
 	if err != nil {
 		return nil, err
 	}
-	return &PiholeProvider{api: api, apiVersion: cfg.APIVersion}, nil
+	return &PiholeProvider{api: api}, nil
 }
 
 // Records implements Provider, populating a slice of endpoints from
@@ -132,17 +115,14 @@ func (p *PiholeProvider) ApplyChanges(ctx context.Context, changes *plan.Changes
 	for _, ep := range changes.UpdateNew {
 		key := piholeEntryKey{ep.DNSName, ep.RecordType}
 
-		// If the API version is 6, we need to handle multiple targets for the same DNS name.
-		if p.apiVersion == "6" {
-			if existing, ok := updateNew[key]; ok {
-				existing.Targets = append(existing.Targets, ep.Targets...)
+		if existing, ok := updateNew[key]; ok {
+			existing.Targets = append(existing.Targets, ep.Targets...)
 
-				// Deduplicate targets
-				slices.Sort(existing.Targets)
-				existing.Targets = slices.Compact(existing.Targets)
+			// Deduplicate targets
+			slices.Sort(existing.Targets)
+			existing.Targets = slices.Compact(existing.Targets)
 
-				ep = existing
-			}
+			ep = existing
 		}
 		updateNew[key] = ep
 	}
@@ -151,18 +131,9 @@ func (p *PiholeProvider) ApplyChanges(ctx context.Context, changes *plan.Changes
 		// Check if this existing entry has an exact match for an updated entry and skip it if so.
 		key := piholeEntryKey{ep.DNSName, ep.RecordType}
 		if newRecord := updateNew[key]; newRecord != nil {
-			// If the API version is 6, we need to handle multiple targets for the same DNS name.
-			if p.apiVersion == "6" {
-				if cmp.Diff(ep.Targets, newRecord.Targets) == "" {
-					delete(updateNew, key)
-					continue
-				}
-			} else {
-				// For API version <= 5, we only check the first target.
-				if newRecord.Targets[0] == ep.Targets[0] {
-					delete(updateNew, key)
-					continue
-				}
+			if cmp.Diff(ep.Targets, newRecord.Targets) == "" {
+				delete(updateNew, key)
+				continue
 			}
 
 			if err := p.api.deleteRecord(ctx, ep); err != nil {
